Fail with an error string when species doesn't exist.
Instead of dealing with species nullability by checking for null, we wrap
it in an Option type, which indicates if the value exists or not.

Instead of throwing when the species is not found, we use an Either
type (also known as Result), which encapsulates the idea
of success and failure.